Made Program class public to enable unit testing.  Added a reference to the main project into the test project.  Made items list protected internal.  I added protected to allow the unit tests to access this member.  I added internal because that is what it was before by default and I want to limit what I am doing in each refactoring step.
Made the list and the UpdateQuality method static so that I don't run into conflicts between instance and non-instance methods.  The client allowed that but was silent on changing the main method.  The Program instance inside the main method isn't ever in a scope such that I can access it from a unit test.  That program instance was used to run the UpdateQuality instance method.